Move platform-specific arg handling to sys::args
