+++ src/tools/tidy/src/pal.rs
@@ -71,7 +71,6 @@ const EXCEPTION_PATHS: &'static [&'static str] = &[
     "src/libstd/num/f64.rs",
     "src/libstd/thread/local.rs",
     "src/libstd/sys/common/mod.rs",
-    "src/libstd/sys/common/args.rs",
     "src/libstd/sys/common/net.rs",
     "src/libstd/sys/common/util.rs",
     "src/libterm", // Not sure how to make this crate portable, but test needs it
